requirements.txt
pillow
pytesseract
pyautogui
opencv-python
numpy
mouse
